modify after code

+++ go/sqlboiler-tutorial/mysql/thelper/after.go
@@ -2,6 +2,7 @@ package thelper
 
 import (
 	"fmt"
+	"log"
 	"sqlboiler-tutorial-mysql/db"
 	"testing"
 
@@ -16,19 +17,51 @@ func FinalizeTest(t *testing.T) {
 
 func truncateTable() {
 	d := db.GetDB()
-	tableNames := GetTableNames()
+	tableNames := getTableNames()
 
 	for _, name := range tableNames {
-		c := fmt.Sprintf("TRUNCATE TABLE %s CASCADE", name)
-		queries.Raw(c).Exec(d)
+		if notContains(getExcludeTables(), name) {
+			c := fmt.Sprintf("TRUNCATE TABLE %s CASCADE", name)
+			queries.Raw(c).Exec(d)
+		}
 	}
 
 	db.Close()
 }
 
-func GetTableNames() []string {
-	// 現状ホワイトリスト形式で対象のテーブルを列挙しているが、自動的に取得できるようにしたい。。。
-	// mysql だと最初から作成されているマスター的なテーブルがあるからそこも削除されないように。
-	tableNames := []string{"users", "posts"}
+func getExcludeTables() []string {
+	return []string{"schema_migrations"}
+}
+
+func getTableNames() []string {
+	d := db.GetDB()
+
+	rows, err := queries.Raw("SHOW TABLES;").Query(d)
+	if err != nil {
+		fmt.Println(err)
+		return nil
+	}
+	defer rows.Close()
+
+	var tableNames []string
+	for rows.Next() {
+		var tableName string
+		err := rows.Scan(&tableName)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		tableNames = append(tableNames, tableName)
+	}
+
 	return tableNames
+}
+
+func notContains(s []string, e string) bool {
+	for _, v := range s {
+		if e == v {
+			return false
+		}
+	}
+	return true
 }
